Adding a pattern for using importlib for problematic ports
